Support firing while moving.
Move checking of firing to a separate function. The new function checks
if the new state allows firing. If it does and the fire key has been
pressed, it sets a flag, frame_shooting, to select the firing frame. The
getTexture() function checks the frame_shooting to select the approprite
texture.