Add support for hitboxes for collision detection.
Allow sprites to define hitboxes for collision detection. To check for
collision we first check that the bounding boxes of the sprites
collide. If they don't then the sprites do not collide. A sprite
collides with another sprite if any of its hitboxes collide with any
of the hitboxes of the other sprite. For sprites that do not define
hitboxes, the bounding box is used as the default hitbox.