Use left-looking sprites for all sprites, except the chameleon.
The right-looking sprites can be drawn by using the horizontal flip
option of the SDL rendering function. Chameleons do not have exactly
simmetrical images when they are looking left or right, so we store
all the images.

The Fred sprite has also been modified to be in white color, and color
modulation can be used to change the color into any color we need.