Add support for objects and the scoreboard.
Add objects, and the ability to pick them up. Add support for writing
text in the Spectrum font, and to write numbers in small digits. Draw
the scoreboard and show the bullet count, level, y position in the
map, power indicator, current score and high score.

The map object is not implemented yet.