Set Window position relative to position of Fred.
Simplify Window data structure: use the position of Fred to determine
the map coordinates, in pixels, of the top left corner of the
window. Allow a user offset so that the window can be moved by the
user as necessary.

This change removes the need for a separate moveFrame() function.