Use callback (pull method) to supply audio data to driver
Set up a callback to get the audio data into the audio device. Use our
own circular buffer instead of the internal SDL audio queue.

This solution is not complete. It's missing support for clearing
queues.

Using this mechanism was useful to determine how frequently the
callback gets called. In Android it was determined that the callback
gets called every 40ms. This is quite significant, and produces a
noticeable lag in some sounds. Depending on this frequency it may be
necessary to sync the game loop and video output with the audio
callback.