Refactor Game class to allow rendering with different renderers
SDL textures objects are associated with a given SDL renderer
object. A TextureManager object holds the pointers to all SDL
textures, which are associated with the renderer used to create the
TextureManager object.

We want to be able to use multiple renderers. We want to use one
renderer to display the image on the window, and a different one to
create bitmaps to save into a file. To do that we need to have one
TextureManager per renderer, and we need to allow the Game class to
use different TextureManager objects and renderers. This patch removes
the TextureManager member in the Game class. The TextureManager object
is now passed as an argument to the Game::render() function.

This patch also removes the SoundManager member from the Game
class. It now gets passed as an argument to the
Game::playPendingSounds() function. This is consistent with the notion
that the Game class should hold the state of the game, and should be
agnostic about how sounds are played, or images are rendered.

This patch also modifies the autofred executable to save the game
window in every frame in a separate file.