Expand game menu
Expand in-game menu. When the user presses Esc, or
makes the "BACK" gesture, the game is paused and
the game menu is shown. There are some game
options, as well as some cheats. If a cheat is
enabled during any part of a game, the score will
not be added to the high score list.