Separate in-game menu between config and cheats (POKEs)
Move the cheats to a separate menu within the
in-game option. Make configuration options global:
that is, changing the configuration in the in-game
menu affects the global configuration (instead of
just the current game), and the modified
configuration gets immediately stored.

Cheats get reset to their default values (which
can be set from the command line) at the beginning
of every game.