Fix out-of-bounds object type
We use a uniform distribution to randomly pick the object type. The
range in this distribution exceded the valid values for the object type.
Because of this some objects were being generated with an object type of
Object::Type::COUNT. This was causing out-of-bounds access to some
arrays, giving invalid values to the bounding box of the object. The
effect was that some objects would seem to be large enough to cover the
whole map, which was causing Fred to immediately pick them at the start
of the game. This manifested as the object pick-up sound being played
at the beginning of the game.